added member to container
